Extract abstract hitobject sample test class
